requirements.txt
pynubank
pyopenssl==22.1.0
python-dotenv